Compute and print homologies
How it works:
The zero columns in the sparse matrix correspond to homologies, but some of these are bounded. We use ranks to check whether this cycle is independent of the image space of boundaries. The persistent homologies are output in order of dimension and persistence.

The implementation for homology computation is in filtration.py

The two new functions in sparse_matrix.py are used to find all cycles (regardless of whether they're bounded)

In visualization, we extract the order in which simplices were added, their birth diameters, and persistence values

An example usage is provided in project_first_hack.py